requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.6.5
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Django==3.2.5
django-crontab==0.7.1
djangorestframework==3.12.4
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
psycopg2==2.9.1
pylint==2.9.6
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
regex==2021.7.6
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.0
urllib3==1.26.6
wrapt==1.12.1
